Solution and Project renamed to "KeyboardShortcuts"
