+++ Framework/KeyboardShortcuts/Config.daml
