+++ Framework/KeyboardShortcuts/Images/AddInDesktop32.png
