+++ Framework/KeyboardShortcuts/KeyboardShortcuts.csproj
