+++ Framework/KeyboardShortcuts/KeyboardShortcuts.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Shortcuts", "Shortcuts.csproj", "{8A539A91-7BA0-4AA1-BB02-82B87EB7631C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KeyboardShortcuts", "KeyboardShortcuts.csproj", "{8A539A91-7BA0-4AA1-BB02-82B87EB7631C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
