+++ Framework/KeyboardShortcuts/Panes/SampleDockPane.xaml
@@ -10,12 +10,12 @@
 //   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //   See the License for the specific language governing permissions and
 //   limitations under the License.-->
-<UserControl x:Class="Shortcuts.Panes.SampleDockPaneView"
+<UserControl x:Class="KeyboardShortcuts.Panes.SampleDockPaneView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"   
-             xmlns:ui="clr-namespace:Shortcuts.Panes"
+             xmlns:ui="clr-namespace:KeyboardShortcuts.Panes"
              xmlns:extensions="clr-namespace:ArcGIS.Desktop.Extensions;assembly=ArcGIS.Desktop.Extensions"
              mc:Ignorable="d" 
              d:DesignHeight="300" d:DesignWidth="300"
