+++ Framework/KeyboardShortcuts/Panes/SamplePaneViewModel.cs
@@ -29,7 +29,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Shortcuts.Panes
+namespace KeyboardShortcuts.Panes
 {
   internal class SamplePaneViewModel : ViewStatePane
   {
