+++ Framework/KeyboardShortcuts/Properties/launchSettings.json
