+++ Framework/KeyboardShortcuts/Screenshots/ConditionShortcut.png
