+++ Framework/KeyboardShortcuts/Screenshots/ShortcutDialog.png
